+++ src/prelude.rs
@@ -2,6 +2,7 @@ pub use crate::request::{ContentTypes, Request, RequestTypes};
 
 pub use crate::form_data::FormData;
 
+#[cfg(feature = "multipart")]
 pub use crate::multipart_form::MultipartFormData;
 
 pub use crate::body::Body;